requirements.txt
streamlit
newspaper3k
transformers
torch
lxml
lxml[html_clean]

